Added front end forms
Changed main.py to app.py
Connected front-end and back-end
Added styling
