+++ static/styles.css
@@ -1,9 +1,46 @@
 main {
   margin: 150px;
 }
-html {
-  font-size: 39px;
+body {
+  font-size: larger;
 }
-input {
-  font-size: 18px;
+.body-content {
+  padding: 40px;
+}
+/*
+input[type='submit'] {
+  padding: 10px;
+  background-color: #ff5678;
+  cursor: pointer;
+}
+*/
+/*New*/
+ul {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+  overflow: hidden;
+  border: 1px solid #e7e7e7;
+  background-color: #f3f3f3;
+}
+
+li {
+  float: left;
+}
+
+li a {
+  display: block;
+  color: #666;
+  text-align: center;
+  padding: 14px 16px;
+  text-decoration: none;
+}
+
+li a:hover:not(.active) {
+  background-color: #ddd;
+}
+
+li a.active {
+  color: white;
+  background-color: #ff5678;
 }
